Add repository.{hpp,cpp} translation unit
